rename: mock 폴더명 변경
