+++ src/app/page.tsx
@@ -1,11 +1,11 @@
+import { postList } from '@/__mock__/postList';
 import PostItem from '@/components/common/post-item/PostItem';
-import { postList } from '@/mocks/postList';
 
 export default function Home() {
   return (
     <main className="mt-5 px-4">
       {postList.post.map(({ author, ...post }) => (
-        <PostItem author={author} post={post} />
+        <PostItem key={post.id} author={author} post={post} />
       ))}
     </main>
   );
